# app/swagger/swagger.yaml
swagger: "2.0"
info:
  title: Job Recruitment Backend API
  description: Swagger documentation for Job Recruitment Backend API
  version: 1.0.0
servers:
  - url: http://localhost:80
    description: Local server
paths:
  /applications/:
    post:
      summary: Apply for a job
      description: 지원하기
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            properties:
              user_id:
                type: string
              job_id:
                type: string
          description: 지원 정보 (user_id와 job_id 필수)
      responses:
        201:
          description: 지원 성공
          schema:
            type: object
            properties:
              message:
                type: string
              application_id:
                type: string
        400:
          description: user_id와 job_id는 필수입니다.
    get:
      summary: Get applications
      description: 지원 내역 조회
      parameters:
        - in: query
          name: user_id
          type: string
          required: true
          description: User ID
      responses:
        200:
          description: 지원 내역 조회 성공
          schema:
            type: object
            properties:
              data:
                type: array
                items:
                  type: object
        400:
          description: user_id는 필수입니다.
  /applications/{application_id}:
    delete:
      summary: Cancel application
      description: 지원 취소
      parameters:
        - in: path
          name: application_id
          required: true
          type: string
          description: Application ID
      responses:
        200:
          description: 지원 취소 성공
        404:
          description: 해당 지원 내역을 찾을 수 없습니다.
    patch:
      summary: Update application status
      description: 지원 상태 업데이트
      parameters:
        - in: path
          name: application_id
          required: true
          type: string
          description: Application ID
        - in: body
          name: body
          required: true
          schema:
            type: object
            properties:
              status:
                type: string
          description: 새 상태
      responses:
        200:
          description: 지원 상태 업데이트 성공
        400:
          description: 상태는 필수입니다.
        404:
          description: 해당 지원 내역을 찾을 수 없습니다.
  /auth/register:
    post:
      summary: Register user
      description: 회원 가입
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            properties:
              email:
                type: string
              password:
                type: string
              username:
                type: string
          description: 이메일과 비밀번호는 필수입니다.
      responses:
        201:
          description: 회원 가입 성공
          schema:
            type: object
            properties:
              message:
                type: string
              user_id:
                type: string
        400:
          description: 이메일과 비밀번호는 필수입니다.
  /auth/login:
    post:
      summary: Login user
      description: 로그인
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            properties:
              email:
                type: string
              password:
                type: string
          description: 이메일과 비밀번호
      responses:
        200:
          description: 로그인 성공
          schema:
            type: object
            properties:
              message:
                type: string
              token:
                type: string
        401:
          description: 이메일 또는 비밀번호가 잘못되었습니다.
  /auth/profile:
    put:
      summary: Update user profile
      description: 회원 정보 수정
      parameters:
        - in: header
          name: Authorization
          required: true
          type: string
          description: Bearer token
        - in: body
          name: body
          required: true
          schema:
            type: object
            properties:
              email:
                type: string
              username:
                type: string
          description: 수정할 정보
      responses:
        200:
          description: 회원 정보 수정 성공
          schema:
            type: object
            properties:
              message:
                type: string
              updated:
                type: boolean
        401:
          description: 유효하지 않은 토큰입니다.
  /auth/delete:
    delete:
      summary: Delete user
      description: 회원 탈퇴
      parameters:
        - in: header
          name: Authorization
          required: true
          type: string
          description: Bearer token
      responses:
        200:
          description: 회원 탈퇴 성공
        401:
          description: 유효하지 않은 토큰입니다.
  /bookmarks/:
    post:
      summary: Add a bookmark
      description: 북마크 추가
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            properties:
              user_id:
                type: string
              job_id:
                type: string
          description: 북마크 정보
      responses:
        201:
          description: 북마크 추가 성공
          schema:
            type: object
            properties:
              message:
                type: string
              bookmark_id:
                type: string
    get:
      summary: Get bookmarks
      description: 북마크 조회
      parameters:
        - in: query
          name: user_id
          type: string
          required: true
          description: User ID
      responses:
        200:
          description: 북마크 조회 성공
          schema:
            type: object
            properties:
              data:
                type: array
                items:
                  type: object
  /bookmarks/{bookmark_id}:
    delete:
      summary: Delete bookmark
      description: 북마크 삭제
      parameters:
        - in: path
          name: bookmark_id
          required: true
          type: string
          description: Bookmark ID
      responses:
        200:
          description: 북마크 삭제 성공
  /bookmarks/filter:
    get:
      summary: Filter bookmarks
      description: 북마크 필터링
      parameters:
        - in: query
          name: filters
          type: object
          description: 필터 조건
      responses:
        200:
          description: 필터링 성공
          schema:
            type: object
            properties:
              data:
                type: array
                items:
                  type: object
  /jobs/:
    get:
      summary: Get all jobs
      description: 채용 공고 조회 (전체)
      parameters:
        - in: query
          name: location
          type: string
          required: false
          description: 위치 필터
        - in: query
          name: 경력
          type: string
          required: false
          description: 경력 필터
        - in: query
          name: company
          type: string
          required: false
          description: 회사명 필터
        - in: query
          name: title
          type: string
          required: false
          description: 제목 필터
        - in: query
          name: deadline
          type: string
          required: false
          description: 마감일 필터
      responses:
        200:
          description: 채용 공고 조회 성공
          schema:
            type: object
            properties:
              data:
                type: array
                items:
                  type: object
    post:
      summary: Add a job posting
      description: 채용 공고 등록
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
          description: 채용 공고 데이터
      responses:
        201:
          description: 채용 공고 등록 성공
          schema:
            type: object
            properties:
              message:
                type: string
              job_id:
                type: string
        400:
          description: 유효한 데이터를 제공하세요.
  /jobs/{job_id}:
    put:
      summary: Update a job posting
      description: 채용 공고 수정
      parameters:
        - in: path
          name: job_id
          required: true
          type: string
          description: Job ID
        - in: body
          name: body
          required: true
          schema:
            type: object
          description: 수정할 데이터
      responses:
        200:
          description: 채용 공고 수정 성공
          schema:
            type: object
            properties:
              message:
                type: string
              updated:
                type: integer
        400:
          description: 유효한 데이터를 제공하세요.
    delete:
      summary: Delete a job posting
      description: 채용 공고 삭제
      parameters:
        - in: path
          name: job_id
          required: true
          type: string
          description: Job ID
      responses:
        200:
          description: 채용 공고 삭제 성공
        404:
          description: 삭제할 데이터를 찾을 수 없습니다.
  /jobs/search:
    get:
      summary: Search for job postings
      description: 채용 공고 검색
      parameters:
        - in: query
          name: query
          type: string
          required: true
          description: 검색어
      responses:
        200:
          description: 채용 공고 검색 성공
          schema:
            type: object
            properties:
              data:
                type: array
                items:
                  type: object
        400:
          description: 검색어를 입력하세요.